requirements.txt
flask
numpy
scikit-learn
pillow
scikit-image
joblib
gunicorn